third_party/WebKit/LayoutTests/http/tests/worklet/dynamic-import.html

<!DOCTYPE html>
<title>Worklet: Dynamic import() on WorkletGlobalScope</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

// This test should not be upstreamed to WPT because this tests console outputs.

console.log('This test logs a result once from each PaintWorkletGlobalScope');
promise_test(() => {
  const kScriptURL = 'resources/dynamic-import-worklet-script.js';
  return CSS.paintWorklet.addModule(kScriptURL)
    .then(result => assert_equals(undefined, result));
}, 'Dynamic import() on WorkletGlobalScope should reject the promise.');

</script>
