third_party/WebKit/LayoutTests/http/tests/worklet/dynamic-import-expected.txt

CONSOLE MESSAGE: line 9: This test logs a result once from each PaintWorkletGlobalScope
CONSOLE ERROR: line 3: TypeError: import() is disallowed on WorkletGlobalScope.
CONSOLE ERROR: line 3: TypeError: import() is disallowed on WorkletGlobalScope.
This is a testharness.js-based test.
PASS Dynamic import() on WorkletGlobalScope should reject the promise.
Harness: the test ran to completion.
